requirements.txt
langchain==0.1.20
langchain-anthropic==0.1.8
crewai==0.28.8
crewai-tools==0.1.6
anthropic>=0.23.0,<1.0.0
python-dotenv==1.0.0
backoff==2.2.1
requests==2.31.0
discord.py==2.3.2
langfuse>=2.0.0